[wasm-ep] remove in-JS VFS tracing support
use the websocket diagnostic server support instead
